Added Exercises: 4.15,4.16,4.17,4.18,4.19,4.20,4.21,4.22,4.23 bloglist expansion, step 1,2,3,4,5,6,7,8,9,10,11

+++ part-4/part4-exercise/controllers/blogRoutes.js
@@ -1,12 +1,37 @@
 
-const blogRoutes = require('express').Router();
-const { findById } = require('../models/blogs');
+const blogRouter = require('express').Router();
 const Blog = require('../models/blogs');
 
-blogRoutes.get('/', async (request, response) => {
+const User = require('../models/users');
+const userExtractor = require('../utils/middleware').userExtractor;
+
+blogRouter.get('/', async (request, response) => {
   try{
-    const blogs = await Blog.find({})
+    const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 ,email: 1})
     response.json(blogs);
+  }
+  catch(err){
+    response.status(400).json(err);
+  }
+});
+
+blogRouter.post('/', userExtractor , async (request, response) => {
+  let body = request.body;
+  const user = request.user;
+
+  const blog = new Blog({
+    title: body.title,
+    author: body.author,
+    likes: body.likes,
+    url : body.url,
+    user: user.id
+  });
+  const userObj = await User.findById(user.id);
+  try{
+    const result = await blog.save()
+    userObj.blogs = userObj.blogs.concat(result._id);
+    await userObj.save();
+    response.status(201).json(result)
   }
   catch(err){
     response.status(400).json(err);
@@ -14,19 +39,8 @@ blogRoutes.get('/', async (request, response) => {
   
 });
 
-blogRoutes.post('/', async (request, response) => {
-  const blog = new Blog(request.body);
-  try{
-    const result = await blog.save()
-    response.status(201).json(result)
-  }
-  catch(err){
-    response.status(400).json(err);
-  }
-});
 
-
-blogRoutes.get('/:id', async (request, response) => {
+blogRouter.get('/:id', async (request, response) => {
   let id = request.params.id;
   try{
     const blog = await Blog.findById(id);
@@ -41,18 +55,24 @@ blogRoutes.get('/:id', async (request, response) => {
   
 });
 
-blogRoutes.delete('/:id', async (request, response) => {
+blogRouter.delete('/:id', userExtractor ,async (request, response) => {
   let id = request.params.id;
+  const user = request.user;
   try{
-    let blog = await Blog.findById(id)
+    let blog = await Blog.findById(id);
     if(blog)
     {
-      await Blog.findByIdAndRemove(id);
-      response.status(204).end();
-      
+      if(blog.user.toString() === user.id.toString())
+      {    
+        await Blog.findByIdAndRemove(id);
+        response.status(200).json({"message":"Blog Deleted!!"});  
+      }
+      else
+        response.status(401).json({"message":"Unauthorized to delete this blog"});
     }
-    else
+    else{
       response.status(404).json({"message":"Not Found"});
+    }
   }
   catch(err){
     response.status(400).json(err);
@@ -61,7 +81,7 @@ blogRoutes.delete('/:id', async (request, response) => {
 });
 
 
-blogRoutes.put('/:id',async (request,response) => {
+blogRouter.put('/:id',async (request,response) => {
   let id = request.params.id;
   try{
     const newUpdatedBlog = await Blog.findByIdAndUpdate(id,request.body)
@@ -73,5 +93,5 @@ blogRoutes.put('/:id',async (request,response) => {
   
 })
 
-module.exports = blogRoutes;
+module.exports = blogRouter;
